+++ pkg/backend/httpstate/backend.go
@@ -36,9 +36,9 @@ import (
 
 	esc_client "github.com/pulumi/esc/cmd/esc/cli/client"
 	"github.com/pulumi/pulumi/pkg/v3/backend"
+	"github.com/pulumi/pulumi/pkg/v3/backend/backenderr"
 	"github.com/pulumi/pulumi/pkg/v3/backend/display"
 	"github.com/pulumi/pulumi/pkg/v3/backend/diy"
-	backendErrors "github.com/pulumi/pulumi/pkg/v3/backend/errors"
 	"github.com/pulumi/pulumi/pkg/v3/backend/httpstate/client"
 	sdkDisplay "github.com/pulumi/pulumi/pkg/v3/display"
 	"github.com/pulumi/pulumi/pkg/v3/engine"
@@ -473,7 +473,7 @@ func (m defaultLoginManager) Login(
 	if !cmdutil.Interactive() {
 		// If interactive mode isn't enabled, the only way to specify a token is through the environment variable.
 		// Fail the attempt to login.
-		return nil, backendErrors.MissingEnvVarForNonInteractiveError{Var: env.AccessToken.Var()}
+		return nil, backenderr.MissingEnvVarForNonInteractiveError{Var: env.AccessToken.Var()}
 	}
 
 	// If no access token is available from the environment, and we are interactive, prompt and offer to
@@ -998,10 +998,10 @@ func (b *cloudBackend) CreateStack(
 			// A 409 error response is returned when per-stack organizations are over their limit,
 			// so we need to look at the message to differentiate.
 			if strings.Contains(errResp.Message, "already exists") {
-				return nil, &backendErrors.StackAlreadyExistsError{StackName: stackID.String()}
+				return nil, &backenderr.StackAlreadyExistsError{StackName: stackID.String()}
 			}
 			if strings.Contains(errResp.Message, "you are using") {
-				return nil, &backendErrors.OverStackLimitError{Message: errResp.Message}
+				return nil, &backenderr.OverStackLimitError{Message: errResp.Message}
 			}
 		}
 		return nil, err
@@ -1359,7 +1359,7 @@ func (b *cloudBackend) createAndStartUpdate(
 	version, token, err := b.client.StartUpdate(ctx, update, tags)
 	if err != nil {
 		if err, ok := err.(*apitype.ErrorResponse); ok && err.Code == 409 {
-			conflict := backendErrors.ConflictingUpdateError{Err: err}
+			conflict := backenderr.ConflictingUpdateError{Err: err}
 			return client.UpdateIdentifier{}, updateMetadata{}, conflict
 		}
 		return client.UpdateIdentifier{}, updateMetadata{}, err
@@ -1683,7 +1683,7 @@ func (b *cloudBackend) GetLatestConfiguration(ctx context.Context,
 	cfg, err := b.client.GetLatestConfiguration(ctx, stackID)
 	switch {
 	case err == client.ErrNoPreviousDeployment:
-		return nil, backendErrors.ErrNoPreviousDeployment
+		return nil, backenderr.ErrNoPreviousDeployment
 	case err != nil:
 		return nil, err
 	default:
